Move converse-caps into the headless build
