+++ src/headless/converse-caps.js
